Adding active outbound peer connections
BGPd will now attempt to actively connect to idle peers

Changes to support this:
- Adding an interval task that checks for idle Peers, attempting to initiate a TCP session
  - Using  library to instantiate a TCPStream with the given BGPd daemon IpAddr (so connections are initiated from a predictable address)
- Since Sessions take ownership of Peers during initial connection, an mpsc channel is used to allow Sessions to send the Peer back to the Peers HashMap when the session is about to end
- PeerConfig now has a  flag. BGPd won't initiate outbound connections if this flag is set to

Other items:
- MessageCodec now uses Capabilities and offers  function
- Adding tests for the  module
- Generally moving towards more idiomatic Rust (as I learn things)

+++ src/main.rs
@@ -56,7 +56,7 @@ fn main() -> Result<()> {
         .unwrap_or("127.0.0.1")
         .parse()
         .expect("Must specify a valid IP Address");
-    let port: u32 = matches
+    let port: u16 = matches
         .value_of("port")
         .unwrap_or("179")
         .parse()
@@ -77,9 +77,6 @@ fn main() -> Result<()> {
     let config = ServerConfig::from_file(&config_path)?;
     debug!("Found {} peers in {}", config.peers.len(), config_path);
 
-    // TODO, setup Server like this
-    // let server = Server(config);
-    // server.serve(addr, port)?;
     serve(addr, port, config)?;
 
     Ok(())
